% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qf_ca_kwic_r.R
\name{mod_qf_ca_kwic_r_ui}
\alias{mod_qf_ca_kwic_r_ui}
\alias{mod_qf_ca_kwic_r_server}
\title{mod_qf_ca_kwic_r_ui and mod_qf_ca_kwic_r_server}
\usage{
mod_qf_ca_kwic_r_ui(id)

mod_qf_ca_kwic_r_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
