% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qf_show_tweets.R
\name{mod_qf_show_tweets_ui}
\alias{mod_qf_show_tweets_ui}
\alias{mod_qf_show_tweets_server}
\title{mod_qf_show_tweets_ui and mod_qf_show_tweets_server}
\usage{
mod_qf_show_tweets_ui(id, tweets_r)

mod_qf_show_tweets_server(input, output, session, tweets)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
