% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise_tweets.R
\name{qf_bind_rows_tweets}
\alias{qf_bind_rows_tweets}
\title{Create a single data frame from a vector of Twitter handles}
\usage{
qf_bind_rows_tweets(users)
}
\arguments{
\item{users}{A character vector of handles or id of Twitter users.}
}
\description{
Create a single data frame from a vector of Twitter handles
}
\examples{

}
