% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise_tweets.R
\name{qf_create_language_list}
\alias{qf_create_language_list}
\title{Create a list of available languages (tipically, to be used in a shiny app)}
\usage{
qf_create_language_list()
}
\description{
Create a list of available languages (tipically, to be used in a shiny app)
}
