% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_emm_newsbrief.R
\name{qf_emm_extract_entities}
\alias{qf_emm_extract_entities}
\title{Extract most frequently used entities in emm newsbrief items}
\usage{
qf_emm_extract_entities(
  language = NULL,
  days = NULL,
  date = NULL,
  n = Inf,
  store = TRUE
)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}

\item{days}{An integer. How many days back in time should be considered, starting from today (0 means today only, 1 means yesterday, etc.)}

\item{date}{Only news downloaded in the given date will be considered. Defaults to all available dates.}

\item{n}{An integer. Number of entities to keep. Defaults to `Inf` (i.e. keeps all entities, ordered by frequency).}
}
\value{
A data.frame (a tibble) with `n` number of rows and two columns, `words` and `n` for number of occurrences.
}
\description{
Extract most frequently used entities in emm newsbrief items
}
\examples{

}
