% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{qf_extract_tweets_by_user}
\alias{qf_extract_tweets_by_user}
\title{Extracts and stores tweets locally}
\usage{
qf_extract_tweets_by_user(users, wait = 1, twitter_token)
}
\arguments{
\item{users}{A character vector of handles or id of Twitter users.}

\item{wait}{An integer, defaults to 1. Seconds to wait between requests to Twitter.}
}
\description{
Extracts and stores tweets locally
}
\examples{

}
