% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qf_find_list_id}
\alias{qf_find_list_id}
\title{Fin the ID of Twitter lists.}
\usage{
qf_find_list_id(slug, owner_user, cache_lists = TRUE,
  overwrite_lists = FALSE, twitter_token = NULL)
}
\arguments{
\item{cache_lists}{Logical, defaults to TRUE. If TRUE, stores locally the lists owned by a given user, in a subfolder `lists_by_user`. If list has already been downloaded, it just loads it. To overwrite, set `overwrite` to TRUE.}

\item{overwrite_lists}{Logical, defaults to FALSE. If TRUE, it overwrites previously downloaded lists.}

\item{twitter_token}{A twitter token generated by rtweet.}

\item{n}{An integer, number of new tweets to request per users, defaults to 200.}
}
\description{
Fin the ID of Twitter lists.
}
\examples{

}
