% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_emm_newsbrief.R
\name{qf_get_emm_newsbrief}
\alias{qf_get_emm_newsbrief}
\title{Get current news in multiple languages from EMM newsbrief}
\usage{
qf_get_emm_newsbrief(
  languages = "all",
  category_id = "rtn",
  keep_xml = TRUE,
  shuffle = TRUE
)
}
\arguments{
\item{languages}{A character vector of one or more languages as two letter codes (e.g. "en"). "all", the dafault, includes all available languages. Current available languages are: `c("ar","bg","cs","da","de","el","en","es", "et","fi","fr","hr", "hu","it", "lt","lv","nl","pl","pt","ro","ru", "sk","sl","sv","sw","tr","zh")`}

\item{category_id}{A character vector of length 1, defaults to "rtn" (default category with all contents). The id for separate categories can be derived from the URL of the given section, e.g. for EC news the id is "ECnews".}

\item{keep_xml}{Logical, defaults to FALSE. If TRUE, removes the xml file downloaded from EMM newsbrief.}

\item{shuffle}{Logical, defaults to TRUE. If TRUE, order in which languages are processed is randomised.}
}
\value{
Nothing, used for its side effects.
}
\description{
Imports and cleans RSS files from [http://emm.newsbrief.eu/]
}
\examples{

}
