% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud_app.R
\name{qf_wordcloud_app}
\alias{qf_wordcloud_app}
\title{Starts wordcloud shiny app}
\usage{
qf_wordcloud_app(shiny_port = 3939, shiny_host = "0.0.0.0")
}
\description{
Starts wordcloud shiny app
}
\examples{
\dontrun{
AnalyseDataset()
}
}
