% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise_tweets.R
\name{qf_bind_rows_tweets}
\alias{qf_bind_rows_tweets}
\title{Create a single data frame from a vector of Twitter handles}
\usage{
qf_bind_rows_tweets(users = NULL, lists = NULL, since = Sys.Date() -
  31, include_rts = FALSE, save = TRUE, twitter_token = NULL)
}
\arguments{
\item{users}{A character vector of handles or id of Twitter users.}

\item{lists}{A character vector of twitter lists, either as IDs or in the form `user/slug``}

\item{since}{A date, expressed in the form Y-M-D (e.g. "2018-12-31")}

\item{save}{Logical, defaults to TRUE. If TRUE, the merged data frame is stored in the following location: `tweets_all/tweets.all.rds`}
}
\description{
Create a single data frame from a vector of Twitter handles
}
\examples{

}
