% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise_tweets.R
\name{qf_create_trending_hashtag_list}
\alias{qf_create_trending_hashtag_list}
\title{Create an ordered list of hashtags (tipically, to be used in a shiny app)}
\usage{
qf_create_trending_hashtag_list()
}
\description{
Create an ordered list of hashtags (tipically, to be used in a shiny app)
}
