% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_get_emm_newsbrief.R
\name{qf_get_emm_newsbrief}
\alias{qf_get_emm_newsbrief}
\title{Get current news in multiple languages from EMM newsbrief}
\usage{
qf_get_emm_newsbrief(
  languages = c("ar", "bg", "cs", "da", "de", "el", "en", "es", "et", "fi", "fr", "hr",
    "hu", "it", "lt", "lv", "nl", "pl", "pt", "ro", "ru", "sk", "sl", "sv", "sw", "tr",
    "zh"),
  shuffle = TRUE
)
}
\arguments{
\item{languages}{A character vector of one or more languages as two letter codes (e.g. "en").}

\item{shuffle}{Logical, defaults to TRUE. If TRUE, order in which languages are processed is randomised.}
}
\value{
Nothing, used for its side effects.
}
\description{
Imports and cleans RSS files from [http://emm.newsbrief.eu/]
}
\examples{

}
